Make help a bit more obvious. Misc tweeks.
Put each yaml in the job_description on its own line so not so
wide of a line. Make delete default None not False in function.

+++ teuthology/beanstalk.py
@@ -19,37 +19,39 @@ def beanstalk_connect(machine_type):
     beanstalk.ignore('default')
     return beanstalk
 
-def walk_jobs(beanstalk, machine_type, delete=False):
+def walk_jobs(beanstalk, machine_type, delete=None):
     curjobs = beanstalk.stats_tube(machine_type)['current-jobs-ready']
     if curjobs != 0:
         x=1
         while x != curjobs:
             x += 1
             job = beanstalk.reserve(timeout=20)
-            if job.body is not None:
-                job_config = yaml.safe_load(job.body)
-                job_name=job_config['name']
-                job_id = job.stats()['id']
-                job_description = job_config['description']
-                if delete:
-                    if delete in job_name:
-                        print 'Deleted Job: {job_id} from queue. Name: {job_name}'.format(
-                        job_id = job_id,
-                        job_name = job_name,
-                        )
-                        job.delete()
+            if job is not None:
+                if job.body is not None:
+                    job_config = yaml.safe_load(job.body)
+                    job_name=job_config['name']
+                    job_id = job.stats()['id']
+                    job_description = job_config['description']
+                    if delete:
+                        if delete in job_name:
+                            print 'Deleted Job: {job_id} from queue. Name: {job_name}'.format(
+                            job_id = job_id,
+                            job_name = job_name,
+                            )
+                            job.delete()
+                        else:
+                            print "Searching queue... Checked " + str(x) + "/" + str(curjobs)," Jobs\r",
                     else:
-                        print "Searching queue... Checked " + str(x) + "/" + str(curjobs)," Jobs\r",
-                else:
-                    print 'Job: {x}/{curjobs} ID: {job_id} Name: {job_name}'.format(
-                    x=x,
-                    curjobs=curjobs,
-                    job_id=job_id,
-                    job_name=job_name,
-                    )
-                    print '\t {desc}'.format(desc=job_description)
-
-        print "\nDone"
+                        print 'Job: {x}/{curjobs} ID: {job_id} Name: {job_name}'.format(
+                        x=x,
+                        curjobs=curjobs,
+                        job_id=job_id,
+                        job_name=job_name,
+                        )
+                        if job_description:
+                            for desc in job_description.split():
+                                print '\t {desc}'.format(desc=desc)
+        log.info("Finished checking Beanstalk Queue.")
     else:
         log.info('No jobs in Beanstalk Queue')
